Add Sales Summary feature in Charts section
This commit includes the addition of the Sales Summary feature to visualize various sales data in the "Charts" section of the application. New functionality includes the ability to filter data by employee and product and to select date ranges for review. A dedicated component was created for this purpose, as well as the necessary requests to populate the aforementioned options. The navigation menu was also updated to include a link to this new feature.