Implement password update and persist user session
The main features in this commit are the implementation of password update logic in the 'UpdatePasswordDialog' component and the addition of persistent user sessions using local storage. Additionally, changes were made to adapt to DateTime usage instead of DateOnly in various components and repositories. Corrections were also made to improve dialog forms and tables in the front-end.